+++ datafile-app-server/src/test/java/org/onap/dcaegen2/collectors/datafile/configuration/AppConfigTest.java
@@ -1,6 +1,7 @@
 /*-
  * ============LICENSE_START======================================================================
- * Copyright (C) 2018, 2020 NOKIA Intellectual Property, 2018-2019 Nordix Foundation. All rights reserved.
+ * Copyright (C) 2018, 2020-2021 NOKIA Intellectual Property, 2018-2019 Nordix Foundation.
+ * All rights reserved.
  * ===============================================================================================
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -83,12 +84,12 @@ public class AppConfigTest {
             .password("izBJD8nLjawq0HMG") //
             .build();
 
-    private static final ImmutableFtpesConfig CORRECT_FTPES_CONFIGURATION = //
-        new ImmutableFtpesConfig.Builder() //
+    private static final ImmutableCertificateConfig CORRECT_CERTIFICATE_CONFIGURATION = //
+        new ImmutableCertificateConfig.Builder() //
             .keyCert("/src/test/resources/dfc.jks") //
             .keyPasswordPath("/src/test/resources/dfc.jks.pass") //
-            .trustedCa("/src/test/resources/ftp.jks") //
-            .trustedCaPasswordPath("/src/test/resources/ftp.jks.pass") //
+            .trustedCa("/src/test/resources/cert.jks") //
+            .trustedCaPasswordPath("/src/test/resources/cert.jks.pass") //
             .build();
 
     private AppConfig appConfigUnderTest;
@@ -120,9 +121,9 @@ public class AppConfigTest {
         Assertions.assertNotNull(publisherCfg);
         assertThat(publisherCfg).isEqualToComparingFieldByField(CORRECT_PUBLISHER_CONFIG);
 
-        FtpesConfig ftpesConfig = appConfigUnderTest.getFtpesConfiguration();
-        assertThat(ftpesConfig).isNotNull();
-        assertThat(ftpesConfig).isEqualToComparingFieldByField(CORRECT_FTPES_CONFIGURATION);
+        CertificateConfig certificateConfig = appConfigUnderTest.getCertificateConfiguration();
+        assertThat(certificateConfig).isNotNull();
+        assertThat(certificateConfig).isEqualToComparingFieldByField(CORRECT_CERTIFICATE_CONFIGURATION);
     }
 
     @Test
@@ -157,7 +158,7 @@ public class AppConfigTest {
         assertThatThrownBy(() -> appConfigUnderTest.getPublisherConfiguration(CHANGE_IDENTIFIER))
             .hasMessageContaining("No PublishingConfiguration loaded, changeIdentifier: PM_MEAS_FILES");
 
-        Assertions.assertNull(appConfigUnderTest.getFtpesConfiguration());
+        Assertions.assertNull(appConfigUnderTest.getCertificateConfiguration());
     }
 
     @Test
@@ -172,7 +173,7 @@ public class AppConfigTest {
         Assertions.assertNull(appConfigUnderTest.getDmaapConsumerConfiguration());
         assertThatThrownBy(() -> appConfigUnderTest.getPublisherConfiguration(CHANGE_IDENTIFIER))
             .hasMessageContaining(CHANGE_IDENTIFIER);
-        Assertions.assertNull(appConfigUnderTest.getFtpesConfiguration());
+        Assertions.assertNull(appConfigUnderTest.getCertificateConfiguration());
     }
 
     @Test
@@ -190,7 +191,7 @@ public class AppConfigTest {
         Assertions.assertNull(appConfigUnderTest.getDmaapConsumerConfiguration());
         assertThatThrownBy(() -> appConfigUnderTest.getPublisherConfiguration(CHANGE_IDENTIFIER))
             .hasMessageContaining(CHANGE_IDENTIFIER);
-        Assertions.assertNull(appConfigUnderTest.getFtpesConfiguration());
+        Assertions.assertNull(appConfigUnderTest.getCertificateConfiguration());
     }
 
     @Test